+++ components/Input/setup.ts
@@ -1,6 +1,7 @@
 import { applyDefaults, cn, opt, partializeClasses } from '../../src/utils.ts';
 import { iComponent, iFwd } from '../../src/types.ts';
 import { iText } from '../Text/setup.ts';
+import { inputStyles } from '../../src/styles.ts';
 
 export type iInput = iComponent<HTMLInputElement> & {
   label: string;
@@ -33,42 +34,43 @@ export default (props: Partial<iInput>) => {
   const classes = partializeClasses({
     input: opt(
       cn(
-        'comp-input',
-        p.error ? 'comp-input-bg_error' : null,
+        'lbx-input',
+        inputStyles.part.input,
+        p.error ? inputStyles.part.bgError : null,
       ),
       p.class,
       p.nostyle || p.nostyleAll,
     ),
     text: opt(cn('select-none'), text?.class, text?.nostyle || p.nostyleAll),
     label: opt(
-      cn('comp-input_label'),
+      cn('lbx-input-label', inputStyles.part.label),
       label?.class,
       label?.nostyle || p.nostyleAll,
     ),
     error: opt(
-      cn('comp-input_error clr-txt-error'),
+      cn(inputStyles.part.error),
       error?.class,
       error?.nostyle || p.nostyleAll,
     ),
     required: opt(
-      cn('comp-input_required clr-txt-error'),
+      cn(inputStyles.part.required),
       required?.class,
       required?.nostyle || p.nostyleAll,
     ),
     container: opt(
       cn(
-        'comp-input_container',
+        inputStyles.part.container,
         p.type && ['button', 'image', 'reset', 'submit'].includes(p.type)
-          ? 'comp-input_button'
+          ? inputStyles.kind.button
           : p.type && ['radio', 'checkbox'].includes(p.type)
-          ? 'comp-input_bool'
+          ? inputStyles.kind.bool
           : p.type &&
               ['datetime-local', 'date', 'month', 'time', 'week'].includes(
                 p.type,
               )
-          ? 'comp-input_date'
-          : 'comp-input_box',
-        p.maxWidth ? 'comp-input_maxwidth' : null,
+          ? [inputStyles.kind.date, 'lbx-input_box']
+          : [inputStyles.kind.box, 'lbx-input_box'],
+        p.maxWidth ? 'lbx-input_maxwidth' : null,
       ),
       container?.class,
       container?.nostyle || p.nostyleAll,